[MOVE/#1] 파일 폴더링 적용
