+++ app/src/main/java/org/sopt/dosopttemplate/data/dto/request/RequestLoginDto.kt
@@ -1,4 +1,4 @@
-package org.sopt.dosopttemplate.data.dto.follower.retrofit2data
+package org.sopt.dosopttemplate.data.dto.request
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
