+++ app/src/main/java/org/sopt/dosopttemplate/data/dto/request/RequestSignUpDto.kt
@@ -1,13 +1,16 @@
-package org.sopt.dosopttemplate.data.dto.follower.retrofit2data
+package org.sopt.dosopttemplate.data.dto.request
 
 import com.google.gson.annotations.SerializedName
 import kotlinx.serialization.Serializable
 
 @Serializable
 
-data class ResponseInquiryDto(
+data class RequestSignUpDto(
     @SerializedName("username")
     val username: String,
+
+    @SerializedName("password")
+    val password: String,
 
     @SerializedName("nickname")
     val nickname: String,
