+++ app/src/main/java/org/sopt/dosopttemplate/data/dto/response/ResponseInquiryDto.kt
@@ -1,16 +1,13 @@
-package org.sopt.dosopttemplate.data.dto.follower.retrofit2data
+package org.sopt.dosopttemplate.data.dto.response
 
 import com.google.gson.annotations.SerializedName
 import kotlinx.serialization.Serializable
 
 @Serializable
 
-data class RequestSignUpDto(
+data class ResponseInquiryDto(
     @SerializedName("username")
     val username: String,
-
-    @SerializedName("password")
-    val password: String,
 
     @SerializedName("nickname")
     val nickname: String,
