+++ app/src/main/java/org/sopt/dosopttemplate/data/dto/response/ResponseSignUpDto.kt
@@ -1,0 +1,5 @@
+package org.sopt.dosopttemplate.data.dto.response
+
+class ResponseSignUpDto {
+
+}
