+++ app/src/main/java/org/sopt/dosopttemplate/module/AuthService.kt
@@ -1,4 +1,4 @@
-package org.sopt.dosopttemplate
+package org.sopt.dosopttemplate.module
 
 import org.sopt.dosopttemplate.retrofit2data.RequestInquiryDto
 import retrofit2.Call
@@ -6,7 +6,7 @@ import org.sopt.dosopttemplate.retrofit2data.RequestLoginDto
 import org.sopt.dosopttemplate.retrofit2data.RequestSignUpDto
 import org.sopt.dosopttemplate.retrofit2data.ResponseInquiryDto
 import org.sopt.dosopttemplate.retrofit2data.ResponseLoginDto
-import org.sopt.dosopttemplate.retrofit2data.ResponseSignUpDto
+import org.sopt.dosopttemplate.data.dto.response.ResponseSignUpDto
 import retrofit2.http.Body
 import retrofit2.http.POST
