+++ app/src/main/java/org/sopt/dosopttemplate/presentation/SignUpActivity.kt
@@ -11,7 +11,7 @@ import org.sopt.dosopttemplate.module.ServicePool.authService
 //import com.codingmy.sopt_w1_hw1.databinding.ActivitySignupBinding
 import org.sopt.dosopttemplate.databinding.ActivitySignupBinding
 import org.sopt.dosopttemplate.presentation.login.LoginActivity
-import org.sopt.dosopttemplate.data.dto.follower.retrofit2data.RequestSignUpDto
+import org.sopt.dosopttemplate.data.dto.request.RequestSignUpDto
 import org.sopt.dosopttemplate.data.dto.follower.retrofit2data.ResponseSignUpDto
 import retrofit2.Call
 import retrofit2.Response
