+++ app/src/main/java/org/sopt/dosopttemplate/presentation/follower/FollowerAdapter.kt
@@ -1,4 +1,4 @@
-package org.sopt.dosopttemplate
+package org.sopt.dosopttemplate.presentation.follower
 
 import android.content.Context
 import android.view.LayoutInflater
