+++ app/src/main/java/org/sopt/dosopttemplate/presentation/follower/FollowerViewHolder.kt
@@ -1,4 +1,4 @@
-package org.sopt.dosopttemplate
+package org.sopt.dosopttemplate.presentation.follower
 
 import androidx.recyclerview.widget.RecyclerView
 import org.sopt.dosopttemplate.databinding.ItemFollowerBinding
