+++ app/src/main/java/org/sopt/dosopttemplate/presentation/mypage/UserViewHolder.kt
@@ -1,6 +1,7 @@
-package org.sopt.dosopttemplate
+package org.sopt.dosopttemplate.presentation.mypage
 
 import androidx.recyclerview.widget.RecyclerView
+import org.sopt.dosopttemplate.data.dataclass.User
 import org.sopt.dosopttemplate.databinding.ItemUserBinding
 
 class UserViewHolder(private val binding: ItemUserBinding) :
