--- app/src/main/java/org/sopt/dosopttemplate/retrofit2data/ResponseSignUpDto.kt
@@ -1,5 +1,0 @@
-package org.sopt.dosopttemplate.retrofit2data
-
-class ResponseSignUpDto {
-
-}
